requirements.txt
numpy
scipy
matplotlib
seaborn
pyyaml
networkx
tqdm

cupy
# cupy-cuda12x
# argparse